feat(mobile-app): Renamed old files
